Update gallery. bigger time displaying, better placements of eventArea.
I have made a distinction between types of times/events. Or you have a
starting and ending date, or you have a date and a starting and ending
time. The first is in case of for example an exhibition, the latter for
a concert. For addition time data/facts the article will offer space…in
case of an opening, and an exhibition and a finnisage.