Changed the single gallery bar
I thinks it is better to have the controls of the gallery above the
images, in case the pictures are in a different format, the controls
will stay at the same position.
Also there is no need for a black bar. This was only the case in the
index page.